+++ CHANGELOG.md
@@ -6,12 +6,13 @@
 - Placeholder completions
   - Placeholder Support for nodes, completions and hovers
 - Hover and Go To Definition support for Placeholder Segments that use a Mob/Item/Custom Placeholder/Stat Id (for instance, <caster.stat.example> or <placeholder.test>)
-- File Type Recognition, with Nodes, Completions, Hovers and all the good stuff for
+- File Type support (with Nodes, Completions, Hovers and all the good stuff) for
   - RandomSpawns
   - Archetypes
   - Reagents
   - Custom MythicRPG Menus
 - Mob Options, DropOptions and DisplayOptions Completions
+- Mob Variables Completions for the variables set on upstream Templates
   
 #### Commands
 - `MythicScribe.putSelectionInsideInlineMetaskill` command and context menu item to put the selected text inside of an inline metaskill